+++ ios/Classes/vision/ImageLabeler.m
@@ -3,6 +3,7 @@
 #import <MLKitImageLabeling/MLKitImageLabeling.h>
 #import <MLKitImageLabelingCommon/MLKitImageLabelingCommon.h>
 #import <MLKitImageLabelingCustom/MLKitImageLabelingCustom.h>
+#import <MLKitLinkFirebase/MLKitLinkFirebase.h>
 
 #define startImageLabelDetector @"vision#startImageLabelDetector"
 #define closeImageLabelDetector @"vision#closeImageLabelDetector"
@@ -33,8 +34,9 @@
     if ([@"default" isEqualToString:type]) {
         MLKImageLabelerOptions *options = [self getImageLabelerOptions:dictionary];
         labeler = [MLKImageLabeler imageLabelerWithOptions:options];
-    } else if ([@"custom" isEqualToString:type]) {
-        MLKCustomImageLabelerOptions *options = [self getCustomLabelerOptions:dictionary];
+    } else if ([@"customLocal" isEqualToString:type] || [@"customRemote" isEqualToString:type]) {
+        MLKCustomImageLabelerOptions *options = [self getCustomLabelerOptions:dictionary result:result];
+        if (options == NULL) return;
         labeler = [MLKImageLabeler imageLabelerWithOptions:options];
     } else {
         NSString *reason =
@@ -76,16 +78,34 @@
     return options;
 }
 
-- (MLKCustomImageLabelerOptions *)getCustomLabelerOptions:(NSDictionary *)optionsData {
-    NSString *modelType = optionsData[@"customModel"];
-    NSString *path = optionsData[@"path"];
+- (MLKCustomImageLabelerOptions *)getCustomLabelerOptions:(NSDictionary *)optionsData result:(FlutterResult)result {
+    NSNumber *local = optionsData[@"local"];
     NSNumber *conf = optionsData[@"confidenceThreshold"];
-    
-    MLKLocalModel *localModel = [[MLKLocalModel alloc] initWithPath:path];
-    
-    MLKCustomImageLabelerOptions *options = [[MLKCustomImageLabelerOptions alloc] initWithLocalModel:localModel];
+    MLKLocalModel *localModel;
+    MLKCustomImageLabelerOptions *options;
+    if (local.boolValue) {
+        NSString *path = optionsData[@"path"];
+        localModel = [[MLKLocalModel alloc] initWithPath:path];
+        options = [[MLKCustomImageLabelerOptions alloc] initWithLocalModel:localModel];
+    } else {
+        NSString *modelName = optionsData[@"modelName"];
+        MLKFirebaseModelSource *firebaseModelSource = [[MLKFirebaseModelSource alloc] initWithName:modelName];
+        MLKCustomRemoteModel *remoteModel = [[MLKCustomRemoteModel alloc] initWithRemoteModelSource:firebaseModelSource];
+        options = [[MLKCustomImageLabelerOptions alloc] initWithRemoteModel:remoteModel];
+        
+        MLKModelManager *modelManager = [MLKModelManager modelManager];
+        
+        BOOL isModelDownloaded = [modelManager isModelDownloaded:remoteModel];
+        
+        if (!isModelDownloaded) {
+            FlutterError *error = [FlutterError errorWithCode:@"Error Model has not been downloaded yet"
+                                                      message:@"Model has not been downloaded yet"
+                                                      details:@"Model has not been downloaded yet"];
+            result(error);
+            return NULL;
+        }
+    }
     options.confidenceThreshold = conf;
-    
     return options;
 }